requirements.txt
llama-index==0.11.18
llama-index-llms-openai==0.2.15
llama-index-llms-ollama==0.3.4
llama-index-embeddings-openai==0.2.5
llama-index-embeddings-ollama==0.3.1
llama-index-embeddings-huggingface==0.3.1
llama-index-embeddings-fastembed==0.2.0
llama-index-vector-stores-qdrant==0.3.2
qdrant-client==1.12.0
python-dotenv==1.0.1
openai==1.52.0
openai-whisper==20240930
pytubefix==8.8.4
moviepy==1.0.3
fastapi==0.115.6
pydantic==2.10.4
uvicorn==0.34.0